[#18] fix: EducationData 패키지명 변경
