+++ src/main/java/kr/sparta/rchive/domain/educationData/enums/EducationTypeEnum.java
@@ -1,4 +1,4 @@
-package kr.sparta.rchive.domain.educationdata.enums;
+package kr.sparta.rchive.domain.educationData.enums;
 
 public enum EducationTypeEnum {
     Sparta_Lecture,
